// nest-api/src/providers/via-cep/via-cep.types.ts
export type TViaCepResponse = {
  cep: string;
  logradouro: string;
  complemento: string;
  unidade: string;
  bairro: string;
  localidade: string;
  uf: string;
  estado: string;
  regiao: string;
  ibge: string;
  gia: string;
  ddd: string;
  siafi: string;
};

export type TViaCepErrorResponse = {
  erro: boolean;
};